@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
     <LangVersion>latest</LangVersion>
 
     <IsPackable>false</IsPackable>
@@ -9,7 +9,7 @@
 
   <ItemGroup>
     <PackageReference Include="Bogus" Version="23.0.2" />
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Include="Moq" Version="4.10.1" />
     <PackageReference Include="Roslynator.Analyzers" Version="2.0.0">
@@ -24,7 +24,9 @@
     </PackageReference>
     <PackageReference Include="FluentAssertions" Version="5.6.0" />
     <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
-    
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="2.2.3"/>
+
+
   </ItemGroup>
 
   <ItemGroup>
