@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
     <LangVersion>latest</LangVersion>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <UserSecretsId>f7d51143-5f51-4aab-97ae-a4bff6b5ca7f</UserSecretsId>
@@ -28,10 +28,6 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="3.0.0" />
     <PackageReference Include="FluentValidation.AspNetCore" Version="8.0.0" />
   </ItemGroup>
-
-  <!--<ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.4" />
-  </ItemGroup>-->
 
   <ItemGroup>
     <ProjectReference Include="..\..\MedEasy.Core\MedEasy.Core.csproj" />
